+++ src/main/kotlin/io/github/t45k/nil/usecase/JavaPreprocess.kt
@@ -1,8 +1,7 @@
-package io.github.t45k.nil.core
+package io.github.t45k.nil.usecase
 
 import io.github.t45k.nil.NILConfig
 import io.github.t45k.nil.entity.CodeBlock
-import io.github.t45k.nil.tokenizer.SymbolSeparator
 import io.reactivex.rxjava3.core.Flowable
 import io.reactivex.rxjava3.kotlin.toFlowable
 import java.io.File
