+++ src/main/kotlin/io/github/t45k/nil/usecase/Location.kt
@@ -1,4 +1,4 @@
-package io.github.t45k.nil.core
+package io.github.t45k.nil.usecase
 
 import io.github.t45k.nil.NILConfig
 import io.github.t45k.nil.entity.NGrams
